iterate on IntUiBridgeLazyTree
